Build fix.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@1669 e19a5227-2556-0410-816f-a35683127991

+++ apps/zxtune-qt/supp/options.cpp
@@ -86,7 +86,7 @@ namespace
     virtual void SetIntValue(const Parameters::NameType& name, Parameters::IntType val)
     {
       Value value(Storage, name);
-      value.Set(QVariant(val));
+      value.Set(QVariant(qlonglong(val)));
     }
 
     virtual void SetStringValue(const Parameters::NameType& name, const Parameters::StringType& val)
